+++ src/main/java/pl/smsapi/api/action/hlr/HlrFactory.java
@@ -1,0 +1,20 @@
+package pl.smsapi.api.action.hlr;
+
+import pl.smsapi.Client;
+import pl.smsapi.api.ActionFactory;
+import pl.smsapi.proxy.Proxy;
+
+public class HlrFactory extends ActionFactory {
+
+    public HlrFactory(Client client, Proxy proxy) {
+        super(client, proxy);
+    }
+
+    public HLRCheckNumber actionCheckNumber(String phoneNumber) {
+        HLRCheckNumber action = new HLRCheckNumber(phoneNumber);
+        action.client(client);
+        action.proxy(proxy);
+
+        return action;
+    }
+}