+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - `pl.smsapi.exception.SmsapiErrorException` to handle API error responses
 - `date_sent` to SMS/MMS send action responses
 - `time_restriction` parameter for SMS send action
+- HLR action factory, `HlrFactory`
+- HLR with IDX support
 
 ### Changed
 - `pl.smsapi.api.UserFactory.actionAdd` marked as deprecated, use `pl.smsapi.api.action.subusers.SubusersFactory.actionAdd` instead
@@ -38,6 +40,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - `pl.smsapi.api.action.sms.SMSGet` without parameters marked as deprecated
 - `pl.smsapi.api.SmsFactory.actionDelete` without parameters marked as deprecated
 - `pl.smsapi.api.action.sms.SMSDelete` without parameters marked as deprecated
+- `pl.smsapi.api.action.hlr.HLRCheckNumber` without parameters marked as deprecated
 
 ### Removed
 - legacy `phonebook.do` contacts API support